Change type price db

+++ priv/repo/migrations/20230831081745_create_products.exs
@@ -5,7 +5,7 @@ defmodule ElixirLearnPhoenix.Repo.Migrations.CreateProducts do
     create table(:products) do
       add :name, :string
       add :thumbnail_url, :string
-      add :price, :integer
+      add :price, :bigint
       add :sold, :integer
       add :rating, :string
       timestamps(default: fragment("NOW()"))
